Extract a generic DiagnosticsFilterPicker with an extra-filters slot
Split the filter picker into a generic base (shared shell, entity-type
group, personal-collections toggle, and an extraFilters slot mirroring
DiagnosticsSidebar's extraInfo) and a thin StaleContentFilterPicker
wrapper that pins the stale entity-type vocabulary.
