+++ enterprise/frontend/src/metabase-enterprise/monitor/content-diagnostics/components/DiagnosticsFilterPicker/DiagnosticsFilterPicker.tsx
@@ -1,0 +1,108 @@
+import { useDisclosure } from "@mantine/hooks";
+import type { ChangeEvent, ReactNode } from "react";
+import { t } from "ttag";
+
+import {
+  Box,
+  Button,
+  Checkbox,
+  FixedSizeIcon,
+  Indicator,
+  Input,
+  Popover,
+  Stack,
+} from "metabase/ui";
+import type { ContentDiagnosticsFilterType } from "metabase-types/api";
+
+import type { ContentDiagnosticsBaseFilterOptions } from "../types";
+import { getFilterTypeLabel } from "../utils";
+
+type DiagnosticsFilterPickerProps<
+  T extends ContentDiagnosticsBaseFilterOptions,
+> = {
+  filterOptions: T;
+  availableTypes: ContentDiagnosticsFilterType[];
+  isDisabled?: boolean;
+  hasDefaultOptions?: boolean;
+  extraFilters?: ReactNode;
+  onFilterOptionsChange: (filterOptions: T) => void;
+};
+
+export function DiagnosticsFilterPicker<
+  T extends ContentDiagnosticsBaseFilterOptions,
+>({
+  filterOptions,
+  availableTypes,
+  isDisabled = false,
+  hasDefaultOptions = false,
+  extraFilters,
+  onFilterOptionsChange,
+}: DiagnosticsFilterPickerProps<T>) {
+  const [isOpened, { toggle, close }] = useDisclosure();
+
+  const handleTypesChange = (newValue: string[]) => {
+    const entityTypes = availableTypes.filter((type) =>
+      newValue.includes(type),
+    );
+    onFilterOptionsChange({ ...filterOptions, entityTypes });
+  };
+
+  const handlePersonalCollectionsChange = (
+    event: ChangeEvent<HTMLInputElement>,
+  ) => {
+    onFilterOptionsChange({
+      ...filterOptions,
+      includePersonalCollections: event.target.checked,
+    });
+  };
+
+  return (
+    <Popover opened={isOpened} onDismiss={close}>
+      <Popover.Target>
+        <Indicator size={8} offset={12} disabled={hasDefaultOptions}>
+          <Button
+            leftSection={<FixedSizeIcon name="filter" aria-hidden />}
+            disabled={isDisabled}
+            data-testid="content-diagnostics-filter-button"
+            onClick={toggle}
+          >
+            {t`Filter`}
+          </Button>
+        </Indicator>
+      </Popover.Target>
+      <Popover.Dropdown>
+        <Box w="20rem" p="md">
+          <Stack>
+            {availableTypes.length > 0 && (
+              <Checkbox.Group
+                label={t`Entity type`}
+                value={filterOptions.entityTypes}
+                onChange={handleTypesChange}
+              >
+                <Stack gap="sm" mt="sm">
+                  {availableTypes.map((type) => (
+                    <Checkbox
+                      key={type}
+                      value={type}
+                      label={getFilterTypeLabel(type)}
+                    />
+                  ))}
+                </Stack>
+              </Checkbox.Group>
+            )}
+            {extraFilters}
+            <Input.Wrapper label={t`Location`}>
+              <Stack gap="sm" mt="sm">
+                <Checkbox
+                  label={t`Include items in personal collections`}
+                  checked={filterOptions.includePersonalCollections}
+                  onChange={handlePersonalCollectionsChange}
+                />
+              </Stack>
+            </Input.Wrapper>
+          </Stack>
+        </Box>
+      </Popover.Dropdown>
+    </Popover>
+  );
+}
