+++ enterprise/frontend/src/metabase-enterprise/monitor/content-diagnostics/components/DiagnosticsFilterPicker/index.ts
@@ -1,0 +1,1 @@
+export * from "./DiagnosticsFilterPicker";
