+++ enterprise/frontend/src/metabase-enterprise/monitor/content-diagnostics/components/types.ts
@@ -1,9 +1,11 @@
 import type { ContentDiagnosticsFilterType } from "metabase-types/api";
 
-export type StaleContentFilterOptions = {
+export type ContentDiagnosticsBaseFilterOptions = {
   entityTypes: ContentDiagnosticsFilterType[];
   includePersonalCollections: boolean;
 };
+
+export type StaleContentFilterOptions = ContentDiagnosticsBaseFilterOptions;
 
 export type ContentDiagnosticsParamsOptions = {
   withSetLastUsedParams?: boolean;
